+++ packages/js/src/plugins/instrumentModule/methods.ts
@@ -3,8 +3,8 @@ import { Leg, QuoteAsset, legBeet } from '@convergence-rfq/rfq';
 import { AccountMeta } from '@solana/web3.js';
 import { createSerializerFromFixableBeetArgsStruct } from '../../types';
 import { addDecimals } from '../../utils/conversions';
+import { toSolitaSide } from '../rfqModule/models/ResponseSide';
 import { LegInstrument, QuoteInstrument } from './types';
-import { toSolitaSide } from "../rfqModule/models/ResponseSide";
 
 export function toLeg(legInstrument: LegInstrument): Leg {
   return {
@@ -16,7 +16,7 @@ export function toLeg(legInstrument: LegInstrument): Leg {
       legInstrument.getDecimals()
     ),
     instrumentDecimals: legInstrument.getDecimals(),
-  side: toSolitaSide(legInstrument.getSide()),
+    side: toSolitaSide(legInstrument.getSide()),
   };
 }