+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @convergence-rfq/cli
 
+## 4.4.5
+
+### Patch Changes
+
+- Updated dependencies [30273e1b]
+  - @convergence-rfq/sdk@4.4.5
+
+## 4.4.4
+
+### Patch Changes
+
+- Updated dependencies [71ed402e]
+  - @convergence-rfq/sdk@4.4.4
+
 ## 4.4.3
 
 ### Patch Changes