+++ packages/js/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @convergence-rfq/sdk
 
+## 4.4.5
+
+### Patch Changes
+
+- 30273e1b: Updated Anchor NoopWallet
+
+## 4.4.4
+
+### Patch Changes
+
+- 71ed402e: Use NoopWallet for Anchor
+
 ## 4.4.3
 
 ### Patch Changes